+++ CocoaHTTPServer/Core/Responses/HTTPRedirectResponse.m
@@ -23,7 +23,7 @@ static const int httpLogLevel = HTTP_LOG_LEVEL_OFF; // | HTTP_LOG_FLAG_TRACE;
 	return self;
 }
 
-- (UInt64)contentLength
+- (NSUInteger)contentLength
 {
 	return 0;
 }
